streamer-backend: don't fail if our SSD is absent
When you shut down a Scaleway instance and boot it again later, you
will get a fresh node assigned. This will have an SSD in it, but its
filesystem will be on a partition, which means it can't mount it and
boot will fail. Adding nofail to the fstab entry ignores that.

Note that you *always* have to deploy the streamer-backend role again
if you reboot the machine, else video dumps will be written to the
local disk, which is way too small.