backend: switch to vlc from mod_nginx_rtml
This is needed, because we're moving to two audio streams in the output
(so sreview can then choose which stream to use when rendering, and maybe
end users can switch between internal camera mic and the PA). We're also
switching from flv to ts for the recordings, as flv is the main limiting
factor.

This is still very much work in progress.