video-box, voctop: fix analyzeduration to 2s
I should've read the documentation properly. This is the amount of TIME
(not bytes, as I imagined) that data is read to try to guess the source's
parameters. As our sources have a keyframe per second, 2 seconds is enough
to catch the stream and not introduce 10s latency in the input, not to
mention overrun sproxy's buffer, as we've moved to 3mbps and above.