video-box: read audio from the audio board
Now we have a cable going from the hagibis to the audio mixer, going into
the PC input. So, we can read directly from it and not the hagibis, and
as a bonus, send it to the room.

There's an added script to make sure the PC input is enabled, due to it
being set to muted by default (together with the USB output), and that
these options are persisted to the teensy.